+++ src/Relabel.php
@@ -39,7 +39,7 @@ class Relabel extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.1.0';
+    public $schemaVersion = '1.1.1';
 
     // Static Properties
     // =========================================================================